Create README.md
Are you a tired teacher? Do you need a fool-proof lesson that is prepared in no time and is also printable in pdf? Look no further. Here you will find an online split crossword generator. Enter vocabulary that is generated into two halves of a crossword for pairwork. Simply enter the vocabulary on the website on your coffee break, hit enter, then print and see the magic happen!